+++ version_c/Project1/operations.h
@@ -3,8 +3,8 @@
 #include "constants.h"
 #include <tuple>
 
-float initial_cost(Clustering* clustering, Config* config);
-float cost_modif(Clustering* c, int f, int t, float* point, Config* config);
+double initial_cost(Clustering* clustering, Config* config);
+double cost_modif(Clustering* c, int f, int t, double* point, Config* config);
 void update(Clustering* c, int f, int t, int p, Config* config);
-std::tuple<float*, float*> compute_new_centroids(Clustering const* c, int f, int t, int p, Config* config);
+std::tuple<double*, double*> compute_new_centroids(Clustering const* c, int f, int t, int p, Config* config);
 #endif
