+++ version_c/Project1/utils.h
@@ -3,7 +3,7 @@
 #include <time.h>
 #include "clustering.h"
 
-float dist(float* p1, float* p2, Config* config);
-bool same_points(float* p1, float* p2, Config* config);
-float prandom(float min, float max);
+double dist(double* p1, double* p2, Config* config);
+bool same_points(double* p1, double* p2, Config* config);
+double prandom(double min, double max);
 #endif
